add more update

+++ database/migrations/2022_06_08_062210_create_kriteria_table.php
@@ -18,7 +18,7 @@ class CreateKriteriaTable extends Migration
             $table->string('kode')->unique();
             $table->string('nama');
             $table->enum('atribut',['cost','benefit']);
-            $table->integer('bobot');
+            $table->double('bobot');
         });
     }
